% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_table.R
\name{anova_table}
\alias{anova_table}
\title{Print ANOVA table}
\usage{
anova_table(data, print = TRUE)
}
\arguments{
\item{data}{list output of `cluster_gen`}

\item{print}{if `TRUE`, output will be a list containing estimators; if `FALSE` (default), output are formatted tables of this information}
}
\value{
Printed ANOVA table
}
\description{
Prints Analysis of Variance table for `cluster_gen` output.
}
