% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_cluster_labels.R
\name{attribute_cluster_labels}
\alias{attribute_cluster_labels}
\title{Attribute Labels in Hierarchical Structure}
\usage{
attribute_cluster_labels(n)
}
\arguments{
\item{n}{numeric vector or list}
}
\value{
list containing appropriate labels for the clusters and their respondents
}
\description{
Attributes cluster and respondent labels in the context of `cluster_gen`.
}
\seealso{
cluster_gen
}
