% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Generate analytical parameters}
\usage{
beta_gen(data, MC = FALSE, replications = 100, output_cov = FALSE,
  rename_to_q = FALSE)
}
\arguments{
\item{data}{output from the \code{questionnaire_gen} function for
\code{full_output = TRUE}}

\item{MC}{if \code{TRUE}, perform Monte Carlo simulation to estimate
regression coefficients}

\item{replications}{for \code{MC = TRUE}, this represents the number of Monte
Carlo subsamples calculated.}

\item{output_cov}{if \code{TRUE}, will also output the covariance matrix of
YXW.}

\item{rename_to_q}{if \code{TRUE}, renames the variables from "x" and "w" to
"q".}
}
\value{
By default, this function will output a vector with the regression
  coefficients, including intercept. If \code{MC == TRUE}, the output will
  instead be a matrix comparing the true regression coefficients obtained
  from the covariance matrix with expected values obtained from a Monte Carlo
  procedure, complete with 99% confidence interval.

  If \code{output_cov = TRUE}, the output will be a list with two elements.
  the first one, \code{betas}, will contain the same output described in the
  previous paragraph. The second element, called \code{vcov_YXW}, contains
  the covariance matrix of the regression coefficients.
}
\description{
Generate analytical parameters
}
\details{
The covariance matrix provided must have Y in the first row/column.
}
\examples{

data <- questionnaire_gen(100, family="gaussian", theta = TRUE,
                           full_output = TRUE, n_X = 2, n_W = list(2, 2, 4))
beta_gen(data, MC = TRUE)

}
