% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_design.R
\name{block_design}
\alias{block_design}
\title{Assignment of test items to blocks.}
\usage{
block_design(n_blocks = NULL, item_parameters, item_block_matrix = NULL)
}
\arguments{
\item{n_blocks}{an integer indicating how many blocks to create.}

\item{item_parameters}{a data.frame of item parameters.}

\item{item_block_matrix}{a matrix of indicators to assign items to blocks.}
}
\description{
\code{block_design} creates a length-2 list containing: 
\itemize{
  \item a matrix that identifies which items correspond to which block and
  \item a table of block descriptive statisics.    
}
}
\section{Warning}{

The default item_block_matrix spirals the items across the n_blocks.
The defualt item_block_matrix requires at least 3 blocks. 
If n_blocks is less than 3, item_block_matrix must be specified.
}
\examples{
item_param <- data.frame(item = seq(1:25), b = runif(25, -2, 2))
ib_matrix <- matrix(nrow = 25, ncol = 5, byrow = FALSE, 
  c(1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
    0,0,0,0,1,1,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
    0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,0,0,0,0,0,0,0,0,0,0,
    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,0,0,0,0,
    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1))
block_design(n_blocks = 5, item_parameters = item_param, item_block_matrix = ib_matrix)
block_design(n_blocks = 5, item_parameters = item_param)

}

