% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_gen.R
\name{book_gen}
\alias{book_gen}
\title{Generate the matrix sampling design}
\usage{
book_gen(n_forms, form_length, book_design = NULL)
}
\arguments{
\item{n_forms}{the number of forms used to create the booklets}

\item{form_length}{the number of items per form}

\item{book_design}{a matrix that describes how the test forms should be combined}
}
\value{
A list of vectors containing the item numbers.  Each vector in the list is a booklet.
}
\description{
An internal function used in \code{test_assembly} that assigns item numbers to booklets
}
\details{
If \code{n_forms} is greater than 2 and \code{is.null(book_design)} is true, booklets are built
using a spiraling design. If \code{n_forms} is less than or equal to 2 then \code{book_design} 
must be specified,
}

