% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booklet_sample.R
\name{booklet_sample}
\alias{booklet_sample}
\title{Assignment of test booklets to test takers.}
\usage{
booklet_sample(n_subj, book_item_design, e = 0.1, iter = 20)
}
\arguments{
\item{n_subj}{integer, the number of subjects (test takers).}

\item{book_item_design}{a data.frame containing the items that belong to each 
booklet with booklets as columns and booklet item numers as rows.}

\item{e}{stopping criteria, the difference between the most sample and least sampled booklets.}

\item{iter}{number if iterations to reach e.}
}
\description{
\code{booklet_sample} randomly assigns test booklets across the n_subj test takers.
}
\section{Details}{

If using \code{booklet_sample} in tandem with \code{booklet_design}, \code{book_item_design}
is the the first element of the returned list of \code{block_design}.
}
\examples{
it_bk <- matrix(c(1, 2, 1, 4, 5, 4, 7, 8, 7, 10, 3, 10, 2, 6, 3, 5, 9, 6, 8, 0, 9), 
           ncol = 3, byrow = TRUE)
booklet_sample(n_subj = 10, book_item_design = it_bk)

}

