% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{brr}
\alias{brr}
\title{Generate replicates of a dataset using Balanced Repeated Replication}
\usage{
brr(data, pseudo_strata = ceiling(nrow(data)/2), weight_cols = "none")
}
\arguments{
\item{data}{dataset}

\item{pseudo_strata}{number of pseudo-strata}

\item{weight_cols}{vector of weight columns}
}
\value{
a list containing all the BRR replicates of `data`
}
\description{
Generate replicates of a dataset using Balanced Repeated Replication
}
\seealso{
cluster_estimates jackknife
}
