% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_n_tilde.R
\name{calc_n_tilde}
\alias{calc_n_tilde}
\title{Calculate ñ}
\usage{
calc_n_tilde(M, N, n_j)
}
\arguments{
\item{M}{total number of population (i.e., sum of n_j over all j)}

\item{N}{number of each class j}

\item{n_j}{vector with size of each class j}
}
\description{
Calculates n tilde
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1999). Multilevel Analysis. Sage Publications.
}
\seealso{
?lsasim:::summary.lsasimcluster
}
