% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_se_rho.R
\name{calc_se_rho}
\alias{calc_se_rho}
\title{Calculate Standard Error of Intraclass Correlation}
\usage{
calc_se_rho(rho, n_j, N)
}
\arguments{
\item{rho}{intraclass correlation}

\item{n_j}{number of elements in class j}

\item{N}{number of classes j}
}
\description{
Calculate Standard Error of Intraclass Correlation
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1999). Multilevel Analysis. Sage Publications.
}
\seealso{
anova.lsasimcluster
}
