% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_var_between.R
\name{calc_var_between}
\alias{calc_var_between}
\title{Calculate variance between classes}
\usage{
calc_var_between(n_j, y_bar_j, y_bar, n_tilde, N)
}
\arguments{
\item{n_j}{number of elements in class j}

\item{y_bar_j}{mean of variable of interest per class j}

\item{y_bar}{mean of variable of interest across classes}

\item{n_tilde}{function of the variance of n_N, M and N. See documentation and code of \code{lsasim:::summary.lsasimcluster} for details}

\item{N}{number of classes j}
}
\description{
Calculate variance between classes
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1999). Multilevel Analysis. Sage Publications.
}
\seealso{
anova.lsasimcluster
}
