% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_condition.R
\name{check_condition}
\alias{check_condition}
\title{Check if an error condition is satisfied}
\usage{
check_condition(condition, message, fatal = TRUE)
}
\arguments{
\item{condition}{logical test which if \code{TRUE} will cause the function to
return an error message}

\item{message}{error message to be displayed if condition is met.}

\item{fatal}{if \code{TRUE}, error message is fatal, i.e., it will abort the
parent function which called \code{check_condition}.}
}
\description{
Check if an error condition is satisfied
}
