% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_N_class.R
\name{check_n_N_class}
\alias{check_n_N_class}
\title{Check class of n or N}
\usage{
check_n_N_class(x)
}
\arguments{
\item{x}{either n or N from `cluster_gen`}
}
\description{
Check the class of an object (usually n and N from `cluster_gen`)
}
\note{
This function is primarily used as a way to simplify the classification of n and N in the `cluster_gen` function.
}
\seealso{
cluster_gen
}
