% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_structure.R
\name{check_valid_structure}
\alias{check_valid_structure}
\title{Check if List is Valid}
\usage{
check_valid_structure(n)
}
\arguments{
\item{n}{list}
}
\value{
Error if the structure is improper. Otherwise, there's no output.
}
\description{
Checks if a list has a proper structure to be transformed into a hierarchical structure
}
\seealso{
check_condition
}
