% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_estimates.R
\name{cluster_estimates}
\alias{cluster_estimates}
\title{Calculate summary statistics for cluster samples}
\usage{
cluster_estimates(data, var_method = "none", print_data = FALSE,
  verbose = FALSE)
}
\arguments{
\item{data}{list of background questionnaire data (typically generated by `cluster_gen`)}

\item{var_method}{if other than `none`, calculates replicate weights. Can be "jackknife", "BRR" or "BRR Fay"}

\item{print_data}{if `TRUE`, includes the input data with replicate weights as an output}

\item{verbose}{if `TRUE`, prints summary}
}
\description{
Takes the output of `cluster_gen` to calculate some summary statistics
}
\details{
Replicate weights can be calculated using the Jackknife for unstratified two-stage sample designs or Balanced Repeated Replication (BRR) with or without Fay's modification.
}
\references{
OECD (2015). Pisa Data Analysis Manual.
    Rust, K. F., & Rao, J. N. K. (1996). Variance estimation for complex surveys using replication techniques. Statistical methods in medical research, 5(3), 283-310.
}
\seealso{
cluster_gen jackknife, jackknife_var
}
