% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen_separate.R
\name{cluster_gen_separate}
\alias{cluster_gen_separate}
\title{Generate cluster samples with individual questionnaires}
\usage{
cluster_gen_separate(n_levels, clusters, n_obs, labels, collapse, n_X, n_W,
  c_mean, ...)
}
\arguments{
\item{n_levels}{number of cluster levels}

\item{clusters}{numeric vector with the number of clusters on each level}

\item{n_obs}{numeric vector with the number of observations in each cluster}

\item{labels}{character vector with the names of each cluster level}

\item{collapse}{if `TRUE`, function output contains only one data frame with all answers}

\item{n_X}{list of `n_X` per cluster level}

\item{n_W}{list of `n_W` per cluster level}

\item{c_mean}{vector of means for the continuous variables or list of vectors for the continuous variables for each level}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\description{
This is a subfunction of `cluster_gen` that performs cluster sampling, with the twist that each cluster level has its own questionnaire.
}
