% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen_together.R
\name{cluster_gen_together}
\alias{cluster_gen_together}
\title{Generate cluster samples with lowest-level questionnaires}
\usage{
cluster_gen_together(n_levels, n_obs, N, sampling_method, cluster_labels,
  resp_labels, collapse, n_X, n_W, c_mean, ...)
}
\arguments{
\item{n_levels}{number of cluster levels}

\item{n_obs}{numeric vector with the number of clusters on each level}

\item{N}{numeric vector with the population size of each level}

\item{sampling_method}{can be "SRS" for Simple Random Sampling or "PPS" for Probabilities Proportional to Size}

\item{cluster_labels}{character vector with the names of each cluster level}

\item{resp_labels}{character vector with the names of the questionnaire respondents on each level}

\item{collapse}{if `TRUE`, function output contains only one data frame with all answers}

\item{n_X}{list of `n_X` per cluster level}

\item{n_W}{list of `n_W` per cluster level}

\item{c_mean}{vector of means for the continuous variables or list of vectors for the continuous variables for each level}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\description{
This is a subfunction of `cluster_gen` that performs cluster sampling where only the lowest-level individuals (e.g. students) fill out questionnaires.
}
