% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_message.R
\name{cluster_message}
\alias{cluster_message}
\title{Print messages about clusters}
\usage{
cluster_message(
  n_obs,
  resp_labels,
  cluster_labels,
  n_levels,
  separate_questionnaires,
  type,
  detail = FALSE
)
}
\arguments{
\item{n_obs}{list with the number of elements per level}

\item{resp_labels}{character vector with the names of the questionnaire respondents on each level}

\item{cluster_labels}{character vector with the names of each cluster level}

\item{n_levels}{number of cluster levels}

\item{separate_questionnaires}{if `TRUE`, each level will have its own questionnaire}

\item{type}{Type of top-level message}

\item{detail}{if `TRUE`, prints further details about each level composition}
}
\value{
Messages.
}
\description{
Prints messages about the cluster scheme before generating questionnaire responses.
}
