% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vector_to_list.R
\name{convert_vector_to_list}
\alias{convert_vector_to_list}
\title{Convert Vector to Expanded List}
\usage{
convert_vector_to_list(x, x_max = x)
}
\arguments{
\item{x}{vector or ranged list to be converted}

\item{x_max}{reference vector or ranged list with max values for x}
}
\value{
expanded/replicated version of x
}
\description{
Converts a vector to list where each element is replicated a certain number of times depending on the previous vector
}
