% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVectorToList.R
\name{convertVectorToList}
\alias{convertVectorToList}
\title{Convert Vector to Expanded List}
\usage{
convertVectorToList(x)
}
\arguments{
\item{x}{vector to ve converted}
}
\value{
expanded/replicated version of x
}
\description{
Converts a vector to list where each element is replicated a certain number of times depending on the previous vector
}
