% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_yfz_gen.R
\name{cov_yfz_gen}
\alias{cov_yfz_gen}
\title{Generate latent regression covariance matrix}
\usage{
cov_yfz_gen(Phi, n_z, sd_z, wcol_Phi, w_names, cor_ptbis)
}
\arguments{
\item{Phi}{latent regression correlation matrix}

\item{n_z}{number of background variables}

\item{sd_z}{standard deviation of background variables}

\item{wcol_Phi}{vector with column numbers of the W variables in Phi}

\item{w_names}{names of W variables}

\item{cor_ptbis}{vector of point-bisserial correlations between Y and F or W}
}
\description{
Generates covariance matrix between Y, F and Z
}
