% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_yxz_gen.R
\name{cov_yxz_gen}
\alias{cov_yxz_gen}
\title{Generate analytical covariance matrix}
\usage{
cov_yxz_gen(vcov_yxw, w_names, Phi, pr_grp_1, n_ind, n_fac, Lambda, var_z)
}
\arguments{
\item{vcov_yxw}{covariance matrix between Y, X and W}

\item{w_names}{name of the W variables}

\item{Phi}{latent regression correlation matrix}

\item{pr_grp_1}{scalar or list of proportions of the first group}

\item{n_ind}{number of indicator variables}

\item{n_fac}{number of factors}

\item{Lambda}{matrix containing the factor loadings}

\item{var_z}{vector of variances of the background variables}
}
\description{
Generate analytical covariance matrix
}
