% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.onAttach}
\alias{.onAttach}
\title{Prints welcome message on package load}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{no idea, but will break devtools::document() if removed}

\item{pkgname}{name of the package}
}
\description{
Prints "This is lsasim <version number>" on package load
}
\note{
This function was adapted from the lavaan package, so credit for it goes to lavaan's creator, Yves Rosseel
}
\references{
Yves Rosseel (2012). lavaan: An R Package for Structural Equation
 Modeling. Journal of Statistical Software, 48(2), 1-36. URL
 http://www.jstatsoft.org/v48/i02/.
}
