% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cat_prop.R
\name{gen_cat_prop}
\alias{gen_cat_prop}
\title{Generates cat_prop for questionnaire_gen}
\usage{
gen_cat_prop(n_X, n_W, n_cat_W)
}
\arguments{
\item{n_X}{number of continuous variables}

\item{n_W}{number of categorical variables}

\item{n_cat_W}{number of categories per categorical variable}
}
\description{
Generates cat_prop for questionnaire_gen
}
