% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_variable_n.R
\name{gen_variable_n}
\alias{gen_variable_n}
\title{Randomly generate the quantity of background variables}
\usage{
gen_variable_n(n_vars, n_X, n_W, theta = FALSE)
}
\arguments{
\item{n_vars}{number of variables in total (\code{n_X + n_W + theta})}

\item{n_X}{number of continuous variables}

\item{n_W}{number of categorical variables}

\item{theta}{number of latent variables}
}
\value{
vector with n_vars, n_X and n_W
}
\description{
Randomly generate the quantity of background variables
}
