% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraclass_cor.R
\name{intraclass_cor}
\alias{intraclass_cor}
\title{Intraclass correlation}
\usage{
intraclass_cor(tau2_hat, sigma2_hat)
}
\arguments{
\item{tau2_hat}{estimate of the true between-class correlation}

\item{sigma2_hat}{estimate of the true within-class correlation}
}
\description{
Calculates the intraclass correlation of clustered data
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1999). Multilevel Analysis. Sage Publications.
}
\seealso{
cluster_gen ?lsasim:::summary.lsasimcluster
}
