% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_gen.R
\name{irt_gen}
\alias{irt_gen}
\title{Simulate item responses from an item response model}
\usage{
irt_gen(theta, a_par = 1, b_par, c_par = 0, D = 1)
}
\arguments{
\item{theta}{numeric ability estimate.}

\item{a_par}{numeric discrimination parameter.}

\item{b_par}{numeric or vector of numerics difficulty parameter(s).}

\item{c_par}{numeric guessing parameter.}

\item{D}{numeric parameter to specify logisitic (1) or normal (1.7).}
}
\description{
Creates a data frame of item parameters.
}
\examples{
irt_gen(theta = 0.2, b_par = 0.6)
irt_gen(theta = 0.2, a_par = 1.15, b_par = 0.6)
irt_gen(theta = 0.2, a_par = 1.15, b_par = 0.6, c_par = 0.2)

}
