% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Generate replicates of a dataset using Jackknife}
\usage{
jackknife(data, weight_cols = "none")
}
\arguments{
\item{data}{dataset}

\item{weight_cols}{vector of weight columns}
}
\value{
a list containing all the Jackknife replicates of `data`
}
\description{
Generate replicates of a dataset using Jackknife
}
\seealso{
cluster_estimates
}
