% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_var.R
\name{jackknife_var}
\alias{jackknife_var}
\title{Jackknife estimate of sampling variance of the mean}
\usage{
jackknife_var(data_whole, data_rep, stat = mean, theta = NULL,
  full_output = FALSE)
}
\arguments{
\item{data_whole}{entire dataset (pre-Jackknife)}

\item{data_rep}{Jackknife list with replications of `data_whole`}

\item{stat}{statistic of interest to calculate (must be a base R function)}

\item{theta}{vector containing the variables of interest}

\item{full_output}{if `TRUE`, returns all intermediate objects created}
}
\description{
Estimates the mean variance for Jackknife replications
}
\details{
`data_rep` can be obtained from
}
\seealso{
jackknife
}
