% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelRespondents.R
\name{labelRespondents}
\alias{labelRespondents}
\title{Label respondents}
\usage{
labelRespondents(n_obs, cluster_labels)
}
\arguments{
\item{n_obs}{list with the number of elements per level}

\item{cluster_labels}{character vector with the names of each cluster level}
}
\value{
Data frame with the combinations of IDs from all levels
}
\description{
This function nerated level label combinations for each respondent
}
