% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_gen.R
\name{lambda_gen}
\alias{lambda_gen}
\title{Randomly generate a matrix of factor loadings}
\usage{
lambda_gen(n_ind, n_fac, limits, row_names, col_names)
}
\arguments{
\item{n_ind}{number of indicators per factor}

\item{n_fac}{number of factors}

\item{limits}{vector with lower and upper limits for the uniformly-generated
Lambdas}

\item{row_names}{vector with row names}

\item{col_names}{vector with col names}
}
\description{
Randomly generate a matrix of factor loadings
}
