% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa2012_cat_pr.R
\docType{data}
\name{pisa2012_cat_pr}
\alias{pisa2012_cat_pr}
\title{Marginal proportions from the PISA 2012 background questionnaire.}
\format{A list of 18 named numeric vectors}
\source{
\url{https://www.oecd.org/pisa/pisaproducts/PISA-2012-technical-report-final.pdf}
}
\usage{
pisa2012_cat_pr
}
\description{
A list containing the marginal proportions for each response category from the PISA 2012 background questionnaire
Elements 1 - 17 are the marginal proportions for the selected items from the background questionnaire.
Element 18 is the marginal proportion for the selected mathematics plausible value.
}
\keyword{datasets}

