% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa2012_cor.R
\docType{data}
\name{pisa2012_cor}
\alias{pisa2012_cor}
\title{Correlation matrix for the backgroun questionnaire}
\format{An 18 by 18 matrix.}
\source{
\url{https://www.oecd.org/pisa/pisaproducts/PISA-2012-technical-report-final.pdf}
}
\usage{
pisa2012_cor
}
\description{
A correlation matrix for the selected background questionnairs and 
mathematics plausible value.
}
\keyword{datasets}

