% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa2012_math_booklet.R
\docType{data}
\name{pisa2012_math_booklet}
\alias{pisa2012_math_booklet}
\title{PISA 2012 mathematics item block - test booklet indicator matrix}
\format{
A data frame with 13 rows and 10 variables:
\describe{
  \item{booklet}{Booklet name.}
  \item{b1}{Indicator specifying those test booklets that use item block 1.}
  \item{b2}{Indicator specifying those test booklets that use item block 2.}
  \item{b3}{Indicator specifying those test booklets that use item block 3.}
  \item{b4}{Indicator specifying those test booklets that use item block 4.}
  \item{b5}{Indicator specifying those test booklets that use item block 5.}
  \item{b6}{Indicator specifying those test booklets that use item block 6.}
  \item{b7}{Indicator specifying those test booklets that use item block 7.}
  \item{b8}{Indicator specifying those test booklets that use item block 8.}
  \item{b9}{Indicator specifying those test booklets that use item block 9.}
}
}
\source{
PISA 2012 Technical Report, Chapter 2: Test Design and Test Development.
        Figure 2.1: Cluster rotation design used to form standard test booklets for PISA 2012. 
        Page 31. 
        \url{https://www.oecd.org/pisa/pisaproducts/PISA-2012-technical-report-final.pdf}
}
\usage{
pisa2012_math_booklet
}
\description{
A dataset containing indicators associating those PISA 2012 mathematics item 
blocks to the PISA 2012 mathematics standard test booklet set.
}
\keyword{datasets}
