% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa2012_math_item_pars.R
\docType{data}
\name{pisa2012_math_item_pars}
\alias{pisa2012_math_item_pars}
\title{Item parameters for 2012 PISA mathematics assessment.}
\format{A data frame with 109 rows and 5 variables:
\describe{
  \item{item_name}{Item name}
  \item{item}{Item number}
  \item{b}{b parameter estimate}
  \item{d1}{d1 parameter estimate (for partial credit items)}
  \item{d2}{d2 parameter estimate (for partial credit items)}
}}
\source{
\url{https://www.oecd.org/pisa/pisaproducts/PISA-2012-technical-report-final.pdf}
}
\usage{
pisa2012_math_item_pars
}
\description{
A dataset containing the estimated item parameters for the PISA 2012 
mathematics assessment.
}
\keyword{datasets}

