% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa2012_q_cormat.R
\docType{data}
\name{pisa2012_q_cormat}
\alias{pisa2012_q_cormat}
\title{Correlation matrix from the PISA 2012 background questionnaire}
\format{
An 19 by 19 matrix.
}
\source{
Raw data can be found at \url{https://www.oecd.org/pisa/pisaproducts/pisa2012database-downloadabledata.htm}
Codebook can be found at \url{https://www.oecd.org/pisa/pisaproducts/PISA12_stu_codebook.pdf}
}
\usage{
pisa2012_q_cormat
}
\description{
A correlation matrix for the selected background questionnaires and
mathematics plausible value.
}
\section{Details}{


A heterogeneous correlation matrix, consisting of polyserial correlations
between numeric and ordinal variables, and polychoric correlations between
ordinal variables.

\tabular{llll}{
 \strong{Row/Col} \tab   \strong{Name}  \tab  \strong{Label}                 \tab  \strong{Type} \cr
  1               \tab   ST93Q01        \tab  Perseverance                   \tab  Ordinal       \cr
  2               \tab   ST93Q03        \tab  Perseverance                   \tab  Ordinal       \cr
  3               \tab   ST93Q04        \tab  Perseverance                   \tab  Ordinal       \cr
  4               \tab   ST93Q06        \tab  Perseverance                   \tab  Ordinal       \cr
  5               \tab   ST93Q07        \tab  Perseverance                   \tab  Ordinal       \cr
  6               \tab   ST94Q05        \tab  Openness for Problem Solving   \tab  Ordinal       \cr
  7               \tab   ST94Q06        \tab  Openness for Problem Solving   \tab  Ordinal       \cr
  8               \tab   ST94Q09        \tab  Openness for Problem Solving   \tab  Ordinal       \cr
  9               \tab   ST94Q10        \tab  Openness for Problem Solving   \tab  Ordinal       \cr
  10              \tab   ST94Q14        \tab  Openness for Problem Solving   \tab  Ordinal       \cr
  11              \tab   ST88Q01        \tab  Attitude toward School         \tab  Ordinal       \cr
  12              \tab   ST88Q02        \tab  Attitude toward School         \tab  Ordinal       \cr
  13              \tab   ST88Q03        \tab  Attitude toward School         \tab  Ordinal       \cr
  14              \tab   ST88Q04        \tab  Attitude toward School         \tab  Ordinal       \cr
  15              \tab   ST89Q02        \tab  Attitude toward School         \tab  Ordinal       \cr
  16              \tab   ST89Q03        \tab  Attitude toward School         \tab  Ordinal       \cr
  17              \tab   ST89Q04        \tab  Attitude toward School         \tab  Ordinal       \cr
  18              \tab   ST89Q05        \tab  Attitude toward School         \tab  Ordinal       \cr
  19              \tab   1PV1MATH       \tab  Mathematics Plausible Value 1  \tab  Continuous    \cr
 }
}

\section{Warning}{


These data are for illustration purposes only.  Handling of missing data may not be suitable for valid inferences.
}

\keyword{datasets}
