% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.R
\name{pluralize}
\alias{pluralize}
\title{Pluralize words}
\usage{
pluralize(word, n = rep(2, length(word)))
}
\arguments{
\item{word}{vector of characters to be pluralized}

\item{n}{vector of number of times each word appears (to determine if the plural or single form will be returned)}
}
\value{
`word`, either pluralized or not (depending on `n`)
}
\description{
Pluralize a word
}
