% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_anova_table.R
\name{print_anova_table}
\alias{print_anova_table}
\title{Print the ANOVA table}
\usage{
print_anova_table(s2_within, s2_between, s2_total, sigma2_hat, tau2_hat,
  rho_hat, se_rho, n_tilde, M, N)
}
\arguments{
\item{s2_within}{Within-class variance}

\item{s2_between}{Between-class variance}

\item{s2_total}{Total variance}

\item{sigma2_hat}{estimate of the true within-class correlation}

\item{tau2_hat}{estimate of the true between-class correlation}

\item{rho_hat}{estimated intraclass correlation}

\item{se_rho}{standard errors of `rho_hat`}

\item{n_tilde}{function of the variance of n_N, M and N. See reference and code within `summarize_clusters` for details}
}
\description{
Print the ANOVA table
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1999). Multilevel Analysis. Sage Publications.
}
\seealso{
anova_table
}
