% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_gen.R
\name{proportion_gen}
\alias{proportion_gen}
\title{Generation of random cumulative proportions}
\usage{
proportion_gen(cat_options, n_cat_options)
}
\arguments{
\item{cat_options}{vector of response types.}

\item{n_cat_options}{vector of number of items of the corresponding response 
type.}
}
\description{
Creates a list of vectors, each containing the randomly generated cumulative 
proportions of a discrete variable.
}
\section{Details}{

\code{cat_options} and \code{n_cat_options} must be the same length.  
\code{cat_options = 1} is a continuous variable.

The result from \code{proportion_gen} can be used directly with the \code{cat_prop}
argument of \code{questionnaire_gen}.
}

\examples{
proportion_gen(cat_options = c(1, 2, 3), n_cat_options = c(2, 2, 2))
proportion_gen(cat_options = c(1, 3), n_cat_options = c(4, 5))

}
