% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_bis_conversion.R
\name{pt_bis_conversion}
\alias{pt_bis_conversion}
\title{Analytical point-biserial conversion}
\usage{
pt_bis_conversion(bis_cor, pr_group1)
}
\arguments{
\item{bis_cor}{biserial correlations}

\item{pr_group1}{probability of group 1}
}
\description{
Analytical point-biserial conversion
}
