% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_gen.R
\name{questionnaire}
\alias{questionnaire}
\title{Generation of discrete and continuous variables}
\usage{
questionnaire(n_obs, cat_prop, cor_matrix, theta = FALSE)
}
\arguments{
\item{n_obs}{number of observations to generate}

\item{cat_prop}{list of cumulative proportions for each item}

\item{cor_matrix}{latent correlation matrix}

\item{theta}{if \code{TRUE} will labeled the first continuous variable 'theta'}
}
\description{
Creates a \code{data.frame} of discrete and continuous variables based on 
an latent correlation matrix and marginal proportions.
}
\details{
\itemize{
  \item \code{cat_prop} is a list where \code{length(cat_prop)} is the number of
        items to be generated.  Each element of the list is a vector containing 
        the marginal cumulative proportions for each category, summing to 1.  For
        continuous items, the associated element in the list should be 1. 
  \item \code{cor_matrix} is correlation matrix that is the same size as 
        \code{length(cat_prop)}.  The correlations related to the correlation 
        between variables on the latent scale. 
  \item \code{theta} is a logical indicator that determines if the first continuous 
        item should be labeled \emph{theta}. If \code{theta = TRUE} but there are no
        continuous variables generated, an error will be returned.    
}
If \code{cat_prop} is a named list, those names will be used as variable names 
for the returned \code{data.frame}.  If \code{cat_prop} is not named and then
generic names will be provided to the variables.
}
\examples{
questionnaire(n = 10, cat_prop = list(c(1), c(.25, .6, 1)), 
             cor_matrix = matrix(c(1, .6, .6, 1), nrow = 2), theta = TRUE)

}

