% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_gen_family.R
\name{questionnaire_gen_family}
\alias{questionnaire_gen_family}
\title{Generation of ordinal and continuous variables}
\usage{
questionnaire_gen_family(n_obs, cat_prop, mean_yw, cov_matrix, family, theta)
}
\arguments{
\item{n_obs}{number of observations to generate.}

\item{cat_prop}{list of cumulative proportions for each item.}

\item{mean_yw}{vector with the means of the latent trait (Y) and the
background variables (W).}

\item{cov_matrix}{covariance matrix. between the latent trait (Y) and the
background variables (W).}

\item{family}{distribution of the background variables. Can be NULL or
'gaussian'.}

\item{theta}{if \code{TRUE} will label the first continuous variable 'theta'.}
}
\description{
Creates a data frame of discrete and continuous variables based on a latent
correlation matrix and marginal proportions.
}
