% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_gen_polychoric.R
\name{questionnaire_gen_polychoric}
\alias{questionnaire_gen_polychoric}
\title{Generation of ordinal and continuous variables}
\usage{
questionnaire_gen_polychoric(n_obs, cat_prop, cor_matrix, c_mean, c_sd, theta)
}
\arguments{
\item{n_obs}{number of observations to generate.}

\item{cat_prop}{list of cumulative proportions for each item.}

\item{cor_matrix}{latent correlation matrix.}

\item{c_mean}{is a vector of population means for each continuous variable.}

\item{c_sd}{is a vector of population standard deviations for each continuous
variable.}

\item{theta}{if \code{TRUE} will label the first continuous variable 'theta'.}
}
\description{
Creates a data frame of discrete and continuous variables based on a latent
correlation matrix and marginal proportions.
}
