% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges}
\alias{ranges}
\title{Defines vector as range}
\usage{
ranges(x, y)
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
`c(x, y)`, but with the "range" class
}
\description{
Redefines the class of a vector as "range"
}
\note{
This function was created to be used as an element in the `N` argument of `cluster_gen`. The name was chosen to avoid conflict with `base::range()`.
}
