% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_final_weights.R
\name{recalc_final_weights}
\alias{recalc_final_weights}
\title{Recalculate final weights}
\usage{
recalc_final_weights(data, w_cols, replicate_weight = 1, reorder = TRUE)
}
\arguments{
\item{data}{dataset}

\item{w_cols}{columns containing the weights}

\item{replicate_weight}{scalar with the replicate weights}

\item{reorder}{if `TRUE`, reorders the dataset so that the replicate weights appear before the final weights}
}
\value{
input data with recalculated final weights, incorporating the replicate weights
}
\description{
Recalculate final weights given the replicate weights
}
