% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_condition_checks.R
\name{run_condition_checks}
\alias{run_condition_checks}
\title{Wrapper-function for check_condition}
\usage{
run_condition_checks(n_cats, n_vars, n_X, n_W, theta, cat_prop, cor_matrix,
  cov_matrix, c_mean, c_sd)
}
\arguments{
\item{n_cats}{vector with number of categories for each categorical variable
(W)}

\item{n_vars}{number of variables (Y, X and W)}

\item{n_X}{number of continuous background variables (X)}

\item{n_W}{number of categorical variables (W)}

\item{theta}{is there a latent variable (Y)?}

\item{cat_prop}{list of vectors with the cumulative proportions of the
background variables}

\item{cor_matrix}{correlation matrix of YXW}

\item{cov_matrix}{covariance matrix of YXW}

\item{c_mean}{vector of means of all variables (YXW)}

\item{c_sd}{vector of standard deviations of all variables (YXW)}
}
\description{
Wrapper-function for check_condition
}
