% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzeropois.R
\name{rzeropois}
\alias{rzeropois}
\title{Generate data from a Zero-truncated Poisson}
\usage{
rzeropois(lambda)
}
\arguments{
\item{lambda}{corresponds to the lambda parameter of a Poisson}
}
\description{
Random generation of one observation of a random variable distributed as a Zero-truncated Poisson
}
\details{
The zero-truncated Poisson (a.k.a. conditional Poisson or positive Poisson) distribution is a discrete probability distribution whose support is the set of positive integers.
}
