% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from.R
\name{sample_from}
\alias{sample_from}
\title{Sample from population structure}
\usage{
sample_from(N, n, labels = names(N), verbose = TRUE)
}
\arguments{
\item{N}{list containing the population sampling structure}

\item{n}{numeric vector with the number of sampled observations (clusters or subjects) on each level}

\item{labels}{character vector with the names of the questionnaire respondents on each level}

\item{verbose}{if `TRUE`, prints output messages}
}
\description{
Generates a sample from a population structure
}
