% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_within_range.R
\name{sample_within_range}
\alias{sample_within_range}
\title{Sample from range}
\usage{
sample_within_range(rg, sample_size = NULL, seed = NULL)
}
\arguments{
\item{rg}{a "range"-class vector}

\item{sample_size}{the size of the sample to be generated}

\item{seed}{pseudo-random number generator seed}
}
\value{
A vector containing the generated sample
}
\description{
Creates a uniformly-distributed sample from a 2-length vector
}
\note{
This function was created primarily to be used to expand an object with the "range" class.
}
