% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Transform regular vector into selection vector}
\usage{
select(...)
}
\arguments{
\item{...}{parameters to be passed to \code{c()}}
}
\value{
same as \code{x}, but with a class attribute that classifies \code{x} as "select"
}
\description{
Attaches a "select" class to a vector
}
\note{
This function was created to be used instead of \code{c()} in the \code{n} argument of \code{cluster_gen}.
}
