% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cat_prop.R
\name{split_cat_prop}
\alias{split_cat_prop}
\title{Split variables in cat_prop}
\usage{
split_cat_prop(cat_prop, keepYX = FALSE)
}
\arguments{
\item{cat_prop}{list corresponding to \code{cat_prop} from
\code{questionnaire_gen}}

\item{keepYX}{if \code{TRUE}, output will be a list separating cat_prop_YX
and cat_prop_W. IF \code{FALSE}, it will be a list with these objects
combined (just like \code{cat_prop})}
}
\description{
Split variables in cat_prop
}
