% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_clusters.R
\name{summarize_clusters}
\alias{summarize_clusters}
\title{Summarizes clusters}
\usage{
summarize_clusters(data)
}
\arguments{
\item{data}{output of `cluster_gen`}
}
\value{
list of summaries
}
\description{
Takes the output of `cluster_gen` and creates summary statistics of the questionnaire variables
}
