% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_assembly.R
\name{test_assembly}
\alias{test_assembly}
\title{Assign test items to subjects}
\usage{
test_assembly(n_subj, n_forms = 1, form_length, book_design = NULL,
  e = 0.1, iter = 20)
}
\arguments{
\item{n_subj}{the number of subjects}

\item{n_forms}{the number of forms used to create the booklets}

\item{form_length}{the number of items per form}

\item{book_design}{a matrix that describes how the test forms should be combined}

\item{e}{stopping criteria for booklet sampling}

\item{iter}{the maximum number of times the booklets will be resampled}
}
\description{
\code{test_assembly} uses a resampling procedure to try to reduce the difference in
booklet coverage.  \code{e} is defined as max(pr) - min(pr) where max(pr) is the proprtion of individuals
who have been assigned to the most used booklet and min(pr) is the proportion of indivudials who
have been assigned the least used booklet.  The default it .1 meaning at most a 10 percent difference,
}

