% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_sample.R
\name{trim_sample}
\alias{trim_sample}
\title{Trim sample}
\usage{
trim_sample(n, N)
}
\arguments{
\item{n}{vector or non-ranged list corresponding to sample structure}

\item{N}{vector or non-ranged list corresponding to population structure}
}
\description{
Makes sure n <= N
}
\seealso{
cluster_gen
}
