% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitelist_message.R
\name{whitelist_message}
\alias{whitelist_message}
\title{Whitelist message}
\usage{
whitelist_message(w)
}
\arguments{
\item{w}{whitelist}
}
\description{
Prints out the sampled elements when cluster_gen is called with select. This function is analogous to cluster_message, but is more proper for random sampling.
}
