% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_irt.R
\name{write.irt}
\alias{write.irt}
\title{Write item response data to a fixed width file.}
\usage{
write.irt(dat, id_var = NULL, theta_var = NULL, group_var = NULL,
  na = 9, ...)
}
\arguments{
\item{dat}{data.frame of item responses.}

\item{id_var}{character name of the id variable in \code{dat}.}

\item{group_var}{character (or vector of) name(s) of grouping variable in \code{dat}.}

\item{na}{integer to replace NA.}

\item{...}{takes arguement from \code{write.fwf}.}
}
\description{
Write item response data to a fixed width file.
}
\examples{
df <- rbind.data.frame(c(1, NA, 1, 0, NA, 0, 0, NA, 1, 0, 1),
                       c(1, 1, NA, 1, 1, NA, 1, 0, NA, 0, 2),
                       c(1, NA, 0, 1, NA, 1, 1, NA, 1, 1, 3),
                       c(1, NA, 0, 0, NA, 1, 1, NA, 0, 0, 4),
                       c(1, NA, 0, 0, NA, 1, 0, NA, 0, 1, 5),
                       c(NA, 0, 0, NA, 0, 0, NA, 0, 0, NA, 6),
                       c(0, 1, NA, 0, 0, NA, 0, 1, NA, 0, 7),
                       c(NA, 1, 0, NA, 1, 1, NA, 1, 1, NA, 8),
                       c(1, 1, NA, 1, 0, NA, 0, 0, NA, 0, 9),
                       c(1, NA, 2, 1, NA, 1, 1, NA, 1, 1, 10),
                       c(NA, 1, 1, NA, 0, 1, NA, 0, 1, NA, 11),
                       c(1, NA, 0, 1, NA, 0, 0, NA, 1, 1, 12),
                       c(NA, 1, 0, NA, 1, 0, NA, 0, 0, NA, 13),
                       c(NA, 1, 0, NA, 2, 0, NA, 1, 0, NA, 14),
                       c(NA, 1, 1, NA, 2, 1, NA, 1, 1, NA, 15),
                       c(NA, 1, 0, NA, 1, 1, NA, 1, 0, NA, 16),
                       c(1, 1, NA, 0, 0, NA, 1, 1, NA, 1, 17),
                       c(1, 0, NA, 0, 1, NA, 1, 0, NA, 0, 18),
                       c(1, 1, NA, 1, 2, NA, 1, 0, NA, 1, 19),
                       c(1, NA, 2, 1, NA, 0, 0, NA, 0, 0, 20))
colnames(df) <- c(paste0("i", seq(10)), "id")
write.irt(dat = df, id_var = "id", sep = "  ")
\dontrun{
  write.irt(df, id_var = "id", na = 8, sep = "  ", file = "my_irt_data.txt")
}
}

