% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_table.R
\name{anova_table}
\alias{anova_table}
\title{Print ANOVA table}
\usage{
anova_table(data, print = TRUE, calc.se = TRUE)
}
\arguments{
\item{data}{list output of `cluster_gen`}

\item{print}{if `TRUE`, output will be a list containing estimators; if `FALSE` (default), output are formatted tables of this information}

\item{calc.se}{if `TRUE`, will try to calculate the standard error of the intreaclass correlation}
}
\value{
Printed ANOVA table
}
\description{
Prints Analysis of Variance table for `cluster_gen` output.
}
\note{
If the rhos for different levels are varied in scale, the generated rho will be less accurate.
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1999). Multilevel Analysis. Sage Publications.
}
