% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_table.R
\name{anova_table}
\alias{anova_table}
\title{Print ANOVA table}
\usage{
anova_table(data)
}
\arguments{
\item{data}{list output of `cluster_gen`}
}
\value{
Printed ANOVA table
}
\description{
Prints Analysis of Variance table for `cluster_gen` output.
}
