% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Generate analytical parameters}
\usage{
beta_gen(data, MC = FALSE, replications = 100)
}
\arguments{
\item{data}{output from the \code{questionnaire_gen} function for
\code{full_output = TRUE}}

\item{MC}{if \code{TRUE}, perform Monte Carlo simulation to estimate
regression coefficients}

\item{replications}{for \code{MC = TRUE}, this represents the number of Monte
Carlo subsamples calculated.}
}
\description{
Generate analytical parameters
}
\details{
The covariance matrix provided must have Y in the first row/column.
}
\examples{

data <- questionnaire_gen(100, family="gaussian", theta = TRUE,
                           full_output = TRUE, n_X = 2, n_W = list(2, 2, 4))
beta_gen(data, MC = TRUE)

}
