% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Generate analytical parameters}
\usage{
beta_gen(data, vcov_yfz, Phi, wcol_Phi, MC = FALSE)
}
\arguments{
\item{data}{output from the \code{questionnaire_gen} function for
\code{full.output = TRUE}}

\item{vcov_yfz}{covariance matrix between Y, F and Z}

\item{Phi}{latent regression correlation matrix}

\item{wcol_Phi}{vector with column numbers of the W variables in Phi}

\item{MC}{if \code{TRUE}, perform Monte Carlo simulation to estimate
regression coefficients}
}
\description{
Generate analytical parameters
}
\examples{
bg <- questionnaire_gen(1000, n_X = 2, n_W = list(2, 2), theta = TRUE,
                        full_output = TRUE)
fit.obs <- lm(theta ~ q1 + q2 + q3 + q4, data = bg$bg)
fit.exp <- beta_gen(bg)
summary(fit.obs)
cbind(obs = fit.obs$coefficients[2:5], beta_hat = fit.exp$beta_hat,
      beta_c = fit.exp$Z0)
}
