% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Generate analytical parameters}
\usage{
<<<<<<< Updated upstream
beta_gen(data, vcov_yfz, Phi, wcol_Phi, prop_groups_1, MC = FALSE,
  replications = 100, analytical = TRUE)
=======
beta_gen(data, vcov_yfz, MC = FALSE, replications = 100)
>>>>>>> Stashed changes
}
\arguments{
\item{data}{output from the \code{questionnaire_gen} function for
\code{full_output = TRUE}}

\item{vcov_yfz}{covariance matrix between Y, F and Z}

\item{MC}{if \code{TRUE}, perform Monte Carlo simulation to estimate
regression coefficients}

\item{replications}{for \code{MC = TRUE}, this represents the number of Monte
Carlo subsamples calculated.}

\item{analytical}{if \code{TRUE}, an analytical solution using the covariance
matrix will be calculated.}
}
\description{
Generate analytical parameters
}
\details{
The covariance matrix provided must have Y in the first row/column.
}
\examples{

# Data containing only continuous variables
data1 <- questionnaire_gen(1000, family="gaussian", theta = TRUE,
                           full_output = TRUE, n_X = 3, n_W = 0)
beta_gen(data1, MC = TRUE)

# Data containing only dichotomous variables
data2 <- questionnaire_gen(1000, family="gaussian", theta = TRUE,
                           full_output = TRUE, n_X = 0, n_W = list(2))
beta_gen(data2, MC = TRUE)

# Data containing only polychotomous variables
data3 <- questionnaire_gen(1000, family="gaussian", theta = TRUE,
                           full_output = TRUE, n_X = 0, n_W = list(3, 5))
\donttest{beta_gen(data3, MC = TRUE)}
}
