% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Generate analytical parameters}
\usage{
beta_gen(vcov_yfz, Phi, wcol_Phi, pr_grp_1)
}
\arguments{
\item{vcov_yfz}{covariance matrix between Y, F and Z}

\item{Phi}{latent regression correlation matrix}

\item{wcol_Phi}{vector with column numbers of the W variables in Phi}

\item{pr_grp_1}{scalar or list of proportions of the first group}
}
\description{
Generate analytical parameters
}
