% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Generate analytical parameters}
\usage{
beta_gen(vcov_yfz, Phi, wcol_Phi, pr_grp_1)
}
\arguments{
\item{vcov_yfz}{covariance matrix between Y, F and Z}

\item{Phi}{latent regression correlation matrix}

\item{wcol_Phi}{vector with column numbers of the W variables in Phi}

\item{pr_grp_1}{scalar or list of proportions of the first group}
}
\description{
Generate analytical parameters
}
\examples{
bg <- questionnaire_gen(100, n_X = 2, n_W = c(2, 2), theta = TRUE,
                        full_output = TRUE)
fit.obs <- lm(theta ~ q1 + q2 + q3 + q4, data = test$bg)
prop.gr.1 <- apply(test$bg[c("q3", "q4")], 2, function(x) prop.table(table(x)))[1, ]
fit.exp <- beta_gen(vcov_yfz = test$cov_matrix, Phi = test$cor_matrix,
                    wcol_Phi = 4:5, pr_grp_1 = prop.gr.1)

cbind(obs = fit.obs$coefficients[2:5], exp = fit.exp$beta_hat)
}
