% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen.R
\name{cluster_gen}
\alias{cluster_gen}
\title{Generate cluster sample}
\usage{
cluster_gen(n, N = 1, cluster_labels = NULL, resp_labels = NULL,
  n_X = NULL, n_W = NULL, c_mean = NULL, c_sd = NULL,
  separate_questionnaires = TRUE, collapse = "none",
  sum_pop = sapply(N, sum), calc_weights = TRUE,
  sampling_method = "mixed", rho = NULL, verbose = TRUE,
  print_pop_structure = verbose, ...)
}
\arguments{
\item{n}{numeric vector with the number of sampled observations (clusters or subjects) on each level}

\item{N}{list of numeric vector with the population size of each *sampled* cluster element on each level}

\item{cluster_labels}{character vector with the names of each cluster level}

\item{resp_labels}{character vector with the names of the questionnaire respondents on each level}

\item{n_X}{list of `n_X` per cluster level}

\item{n_W}{list of `n_W` per cluster level}

\item{c_mean}{vector of means for the continuous variables or list of vectors for the continuous variables for each level}

\item{c_sd}{vector of standard deviations for the continuous variables or list of vectors for the continuous variables for each level}

\item{separate_questionnaires}{if `TRUE`, each level will have its own questionnaire}

\item{collapse}{if `TRUE`, function output contains only one data frame with all answers}

\item{sum_pop}{total population at each level (sampled or not)}

\item{calc_weights}{if `TRUE`, sampling weights are calculated}

\item{sampling_method}{can be "SRS" for Simple Random Sampling or "PPS" for Probabilities Proportional to Size}

\item{rho}{estimated intraclass correlation}

\item{verbose}{if `TRUE`, prints output messages}

\item{print_pop_structure}{if `TRUE`, prints the population hierarchical structure (as long as it differs from the sample structure)}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\value{
list with background questionnaire data, grouped by level or not
}
\description{
Generate cluster sample
}
\details{
This function relies heavily in two subfunctions---`cluster_gen_separate` and `cluster_gen_together`---which can be called independently. This does not make `cluster_gen` a simple wrapper function, as it performs several operations prior to calling its subfunctions, such as randomly generating `n_X` and `n_W` if they are not determined by user.
`n` can have unitary length, in which case all clusters will have the same size.
`N` is *not* the population size across all elements of a level, but the population size for each element of one level.
Regarding the additional parameters to be passed to `questionnaire_gen()`, they can be passed either in the same format as `questionnaire_gen()` or as more complex objects that contain information for each cluster level.
}
\note{
For the purpose of this function, levels are counted starting from the top nesting/clustering level. This means that, for example, schools are the first cluster level, classes are the second, and students are the third and final level. This behavior can be customized by naming the `n` argument or providing custom labels in `cluster_labels` and `resp_labels`.
#TODO: explain the danger of setting c_mean and rho (sigma2 may get really high)

labeling c_mean has no effect, it's for the user.
}
\seealso{
cluster_estimates cluster_gen_separate cluster_gen_together
}
