% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen.R
\name{cluster_gen}
\alias{cluster_gen}
\title{Generate cluster sample}
\usage{
cluster_gen(levels, n_obs = 10, labels = c("class", "school",
  "neighborhood", "city", "region", "country"), collapse = TRUE,
  c_mean = 0, ...)
}
\arguments{
\item{levels}{numeric vector with the number of clusters on each level}

\item{n_obs}{numeric vector with the number of observations in each cluster}

\item{labels}{character vector with the names of each level}

\item{collapse}{if `TRUE`,}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\description{
Generate cluster sample
}
\details{
`n_obs` can have unitary length, in which case all clusters will have the same size.
  Regarding the additional parameters to be passed to `questionnaire_gen()`, they can be passed either in the same format as `questionnaire_gen()` or as more complex objects that contain information for each cluster level.
}
\examples{
cluster_gen(c(1, 2), collapse = FALSE)
}
