% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen.R
\name{cluster_gen}
\alias{cluster_gen}
\title{Generate cluster sample}
\usage{
cluster_gen(clusters, n_obs = 5, labels = c("country", "school",
  "class"), collapse = FALSE, c_mean = 0,
  separate_questionnaires = TRUE, ...)
}
\arguments{
\item{clusters}{numeric vector with the number of clusters on each level}

\item{n_obs}{numeric vector with the number of observations in each cluster}

\item{labels}{character vector with the names of each level}

\item{collapse}{if `TRUE`,}

\item{c_mean}{vector of means for the continuous variables or list of vectors for the continuous variables for each level}

\item{separate_questionnaires}{if `TRUE`, each level will have its own questionnaire}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\description{
Generate cluster sample
}
\details{
`n_obs` can have unitary length, in which case all clusters will have the same size.
  Regarding the additional parameters to be passed to `questionnaire_gen()`, they can be passed either in the same format as `questionnaire_gen()` or as more complex objects that contain information for each cluster level.
}
\examples{
cluster_gen(c(1, 2), collapse = FALSE)
cluster_gen(c(4, 2), n_X = 1, n_W = list(2, 3), theta = TRUE,
            c_mean = list(0, c(0, 10)))
}
