% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen.R
\name{cluster_gen}
\alias{cluster_gen}
\title{Generate cluster sample}
\usage{
cluster_gen(n_obs, cluster_labels = c("country", "school",
  "class")[n_obs], resp_labels = c("principal", "teacher",
  "student")[n_obs], n_X = NULL, n_W = NULL, c_mean = NULL,
  separate_questionnaires = TRUE, collapse = "none", verbose = TRUE,
  ...)
}
\arguments{
\item{n_obs}{numeric vector with the number of observations (clusters or subjects) on each level}

\item{cluster_labels}{character vector with the names of each cluster level}

\item{resp_labels}{character vector with the names of the questionnaire respondents on each level}

\item{n_X}{list of `n_X` per cluster level}

\item{n_W}{list of `n_W` per cluster level}

\item{c_mean}{vector of means for the continuous variables or list of vectors for the continuous variables for each level}

\item{separate_questionnaires}{if `TRUE`, each level will have its own questionnaire}

\item{collapse}{if `TRUE`, function output contains only one data frame with all answers}

\item{verbose}{if `TRUE`, prints output messages}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\description{
Generate cluster sample
}
\details{
This function relies heavily in two subfunctions---`cluster_gen_separate` and `cluster_gen_together`---which can be called independently. This does not make `cluster_gen` a simple wrapper function, as it performs several operations prior to calling its subfunctions, such as randomly generating `n_X` and `n_W` if they are not determined by user.
  `n_obs` can have unitary length, in which case all clusters will have the same size.
  Regarding the additional parameters to be passed to `questionnaire_gen()`, they can be passed either in the same format as `questionnaire_gen()` or as more complex objects that contain information for each cluster level.
}
