% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen_separate.R
\name{cluster_gen_separate}
\alias{cluster_gen_separate}
\title{Generate cluster samples with individual questionnaires}
\usage{
cluster_gen_separate(n_levels, n, N, sum_pop, calc_weights,
  sampling_method, cluster_labels, resp_labels, collapse, n_X, n_W, c_mean,
  verbose, rho, sigma2, ...)
}
\arguments{
\item{n_levels}{number of cluster levels}

\item{n}{numeric vector with the number of sampled observations (clusters or subjects) on each level}

\item{N}{list of numeric vector with the population size of each *sampled* cluster element on each level}

\item{sum_pop}{total population at the lowest level (sampled or not)}

\item{calc_weights}{if `TRUE`, sampling weights are calculated}

\item{sampling_method}{can be "SRS" for Simple Random Sampling or "PPS" for Probabilities Proportional to Size, "mixed" to use SRS for students and PPS otherwise or a vector with the sampling method for each level}

\item{cluster_labels}{character vector with the names of each cluster level}

\item{resp_labels}{character vector with the names of the questionnaire respondents on each level}

\item{collapse}{if `TRUE`, function output contains only one data frame with all answers}

\item{n_X}{list of `n_X` per cluster level}

\item{n_W}{list of `n_W` per cluster level}

\item{c_mean}{vector of means for the continuous variables or list of vectors for the continuous variables for each level}

\item{verbose}{if `TRUE`, prints output messages}

\item{rho}{estimated intraclass correlation}

\item{sigma2}{within-group variance}

\item{...}{Additional parameters to be passed to `questionnaire_gen()`}
}
\description{
This is a subfunction of `cluster_gen` that performs cluster sampling, with the twist that each cluster level has its own questionnaire.
}
\seealso{
cluster_gen cluster_gen_together
}
