% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vector_to_list.R
\name{convert_vector_to_list}
\alias{convert_vector_to_list}
\title{Convert Vector to Expanded List}
\usage{
convert_vector_to_list(x, ref_x = x)
}
\arguments{
\item{x}{vector to be converted}

\item{ref_x}{reference vector}
}
\value{
expanded/replicated version of x
}
\description{
Converts a vector to list where each element is replicated a certain number of times depending on the previous vector
}
