% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gen.R
\name{cor_gen}
\alias{cor_gen}
\title{Generation of random correlation matrix.}
\usage{
cor_gen(n_var)
}
\arguments{
\item{n_var}{integer number of variables.}
}
\description{
Creates a random correlation matrix.
}
\examples{
cor_gen(n_var = 10)

}

