% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_gen.R
\name{cov_gen}
\alias{cov_gen}
\title{Generation of covariance matrices}
\usage{
cov_gen(pr_grp_1 = 0.66, n_fac = 9, n_ind = 4, Lambda_lims = c(0.6,
  0.9), seed = runif(n = 1, max = 1e+08))
}
\arguments{
\item{pr_grp_1}{proportion of observations in group 1}

\item{n_fac}{number of factors}

\item{n_ind}{number of indicators per factor}

\item{Lambda_lims}{vector of lower and upper limits for the factor-loading
matrix}

\item{seed}{sample seed number for the Random Number Generator}
}
\value{
A list containing three covariance matrices: vcov_yxw, vcov_yxz and
  vcov_yfz
}
\description{
Second version of the questionnaire_gen function, developed independently to
maintain functionality of the original function. This function could
eventually be integrated into questionnaire_gen.
}
\examples{
 vcov <- cov_gen()
 str(vcov)
}
