% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_yfz_gen.R
\name{cov_yfz_gen}
\alias{cov_yfz_gen}
\title{Generate latent regression covariance matrix}
\usage{
cov_yfz_gen(n_ind, n_fac, Phi, n_z, sd_z, w_names, pr_grp_1)
}
\arguments{
\item{n_ind}{number of indicator variables}

\item{n_fac}{number of factors}

\item{Phi}{latent regression correlation matrix}

\item{n_z}{number of background variables}

\item{sd_z}{standard deviation of background variables}

\item{w_names}{names of W variables}

\item{pr_grp_1}{scalar or list of proportions of the first group}
}
\description{
Generates covariance matrix between Y, F and Z
}
