% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_yxw_gen.R
\name{cov_yxw_gen}
\alias{cov_yxw_gen}
\title{Setup full YXW covariance matrix}
\usage{
cov_yxw_gen(n_ind, n_z, Phi, n_fac, Lambda)
}
\arguments{
\item{n_ind}{number of indicator variables}

\item{n_z}{number of background variables}

\item{Phi}{latent regression correlation matrix}

\item{n_fac}{number of factor variables}

\item{Lambda}{matrix containing the factor loadings}
}
\description{
Setup full YXW covariance matrix
}
