% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_summary.R
\name{customize_summary}
\alias{customize_summary}
\title{Customize Summary}
\usage{
customize_summary(df_summary, df, numeric_cols, factor_cols, digits = 3)
}
\arguments{
\item{df_summary}{dataframe containing summary statistics}

\item{df}{original data frame}

\item{numeric_cols}{indices of the numeric columns}

\item{factor_cols}{indices of the factor columns}

\item{digits}{controls the number of digits in the output}
}
\description{
Adds standard deviations and removes quantiles from a `summary()` output
}
\seealso{
summary ?lsasim:::summary.lsasimcluster
}
