% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen_auxiliaryFunctions.R
\name{drawClusterStructure}
\alias{drawClusterStructure}
\title{Draw Cluster Structure}
\usage{
drawClusterStructure(n, labels, resp)
}
\arguments{
\item{n}{same from cluster_gen}

\item{labels}{corresponds to cluster_labels from cluster_gen}

\item{resp}{corresponds to resp_labels from cluster_gen}
}
\value{
Prints structure to console.
}
\description{
This function creates a visual representation of the hierarchical structure
}
