% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gen_auxiliaryFunctions.R
\name{select}
\alias{select}
\title{Transform regular vector into selection vector}
\usage{
select(x, ...)
}
\arguments{
\item{x}{vector}
}
\value{
same as `x`, but with a class attribute that classifies `x` as "select"
}
\description{
Attaches a "select" class to a vector
}
