% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_2.R
\name{summary_2}
\alias{summary_2}
\title{Dataset summary statistics}
\usage{
summary_2(data, digits = 3)
}
\arguments{
\item{data}{Data frame}

\item{digits}{number of digits for the output}
}
\description{
Creates summary statistics of a dataset
}
\note{
This function is inspired by base::summary(), but outputs content more relevant to the context of cluster_gen() and summary()
}
\seealso{
summary()
}
